# specs/F04-content-administration/contracts/f04-openapi.yaml
openapi: 3.1.0
info:
  title: Pchinese F04 Content Administration API
  version: 1.0.0-design
  description: >
    Design contract for versioned ADMIN content management. Public read paths
    expose only publishable Free content; administrative mutations use the
    standard success/data/error/meta envelope and require expectedVersion.
    Management responses are typed safe projections, never persistence entities;
    lists are bounded and media intake is server-controlled.
servers:
  - url: /api/v1
...
paths:
  /admin/content/topics:
    get:
      tags: [AdminContent]
      summary: List versioned topic projections for ADMIN
      operationId: listAdminTopics
      parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Size' } ]
      responses:
        '200': { $ref: '#/components/responses/ContentListSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
  /admin/content/lessons:
    get:
      tags: [AdminContent]
      summary: List versioned lesson projections for ADMIN
      operationId: listAdminLessons
      parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Size' } ]
      responses:
        '200': { $ref: '#/components/responses/ContentListSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
  /admin/content/segments:
    get:
      tags: [AdminContent]
      summary: List versioned segment projections for ADMIN
      operationId: listAdminSegments
      parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Size' } ]
      responses:
        '200': { $ref: '#/components/responses/ContentListSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
...
  /admin/content/topics/{topicId}:
    get:
      tags: [AdminContent]
      summary: Reload one safe topic projection for ADMIN
      operationId: getAdminTopic
      parameters: [ { $ref: '#/components/parameters/TopicId' } ]
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/content/lessons/{lessonId}:
    get:
      tags: [AdminContent]
      summary: Reload one safe lesson projection for ADMIN
      operationId: getAdminLesson
      parameters: [ { $ref: '#/components/parameters/LessonId' } ]
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/content/segments/{segmentId}:
    get:
      tags: [AdminContent]
      summary: Reload one safe segment projection for ADMIN
      operationId: getAdminSegment
      parameters: [ { $ref: '#/components/parameters/SegmentId' } ]
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /topics:
    post:
      tags: [AdminContent]
      summary: Create a draft topic
      operationId: createTopic
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TopicCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
  /topics/{topicId}:
    patch:
      tags: [AdminContent]
      summary: Update an existing topic with optimistic versioning
      operationId: updateTopic
      parameters: [ { $ref: '#/components/parameters/TopicId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TopicUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
...
  /lessons:
    post:
      tags: [AdminContent]
      summary: Create a Free lesson
      operationId: createLesson
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LessonCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
  /lessons/{lessonId}:
    patch:
      tags: [AdminContent]
      summary: Update an existing Free lesson
      operationId: updateLesson
      parameters: [ { $ref: '#/components/parameters/LessonId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionedContentRequest' }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
...
  /segments:
    post:
      tags: [AdminContent]
      summary: Create a lesson segment
      operationId: createSegment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SegmentCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
  /segments/{segmentId}:
    patch:
      tags: [AdminContent]
      summary: Update an existing lesson segment with optimistic versioning
      operationId: updateSegment
      parameters: [ { $ref: '#/components/parameters/SegmentId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SegmentUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/ContentSuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/AuthorizationDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StateConflict' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    TopicId: { name: topicId, in: path, required: true, schema: { type: string, format: uuid } }
    LessonId: { name: lessonId, in: path, required: true, schema: { type: string, format: uuid } }
    SegmentId: { name: segmentId, in: path, required: true, schema: { type: string, format: uuid } }
...
    Page: { name: page, in: query, required: false, schema: { type: integer, minimum: 0, default: 0 } }
    Size: { name: size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  responses:
    ContentSuccess:
      description: Latest safe versioned content projection
      content: { application/json: { schema: { $ref: '#/components/schemas/ContentEnvelope' } } }
    ContentListSuccess:
      description: Paginated safe content-management projections
      content: { application/json: { schema: { $ref: '#/components/schemas/ContentListEnvelope' } } }
    ValidationError:
      description: Safe input or lifecycle validation error
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Unauthenticated:
      description: Missing, expired, revoked or invalid access credential
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    AuthorizationDenied:
      description: Caller lacks ADMIN role
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Resource absent or unavailable
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    StateConflict:
      description: Stale expectedVersion or forbidden state transition
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
    TopicCreateRequest:
      type: object
      required: [title, slug, hskLevel, sortOrder]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, pattern: '^[a-z0-9-]+$' }
        summary: { type: string, maxLength: 4000 }
        hskLevel: { type: integer, minimum: 1, maximum: 9 }
        sortOrder: { type: integer, minimum: 0 }
      additionalProperties: false
    TopicUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/TopicCreateRequest'
        - type: object
          required: [expectedVersion]
          properties: { expectedVersion: { type: integer, minimum: 0 } }
    LessonCreateRequest:
      type: object
      required: [topicId, title, slug, hskLevel, sortOrder]
      properties:
        topicId: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, pattern: '^[a-z0-9-]+$' }
        hskLevel: { type: integer, minimum: 1, maximum: 9 }
        sortOrder: { type: integer, minimum: 0 }
        accessLevel: { type: string, const: FREE }
      additionalProperties: false
    VersionedContentRequest:
      type: object
      required: [expectedVersion]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        sortOrder: { type: integer, minimum: 0 }
        expectedVersion: { type: integer, minimum: 0 }
      additionalProperties: false
    SegmentCreateRequest:
      type: object
      required: [lessonId, mediaAssetId, sequenceNo, transcriptHanzi]
      properties:
        lessonId: { type: string, format: uuid }
        mediaAssetId: { type: string, format: uuid }
        sequenceNo: { type: integer, minimum: 1 }
        segmentType: { type: string, enum: [DICTATION, SHADOWING, BOTH, PLAYBACK_ONLY] }
        startMilliseconds: { type: integer, minimum: 0 }
        endMilliseconds: { type: integer, minimum: 0 }
        transcriptHanzi: { type: string, minLength: 1, maxLength: 20000 }
        transcriptPinyin: { type: string, maxLength: 20000 }
        translationVi: { type: string, maxLength: 20000 }
        dictationHint: { type: string, maxLength: 4000 }
      additionalProperties: false
    SegmentUpdateRequest:
      type: object
      required: [expectedVersion]
      properties:
        mediaAssetId: { type: string, format: uuid }
        sequenceNo: { type: integer, minimum: 1 }
        transcriptHanzi: { type: string, maxLength: 20000 }
        transcriptPinyin: { type: string, maxLength: 20000 }
        translationVi: { type: string, maxLength: 20000 }
        dictationHint: { type: string, maxLength: 4000 }
        expectedVersion: { type: integer, minimum: 0 }
      additionalProperties: false
...
    ContentProjection:
      description: Safe typed ADMIN projection. It never includes persistence-only identifiers, actor identifiers, learner data, submitted provider URLs, storage keys, or verification diagnostics.
      type: object
      required: [id, entityType, publicationStatus, version]
      properties:
        id: { type: string, format: uuid }
        entityType: { type: string, enum: [TOPIC, LESSON, SEGMENT, MEDIA] }
        publicationStatus: { type: string, enum: [DRAFT, PUBLISHED, UNPUBLISHED, ARCHIVED, PENDING_SCAN, APPROVED, REJECTED, QUARANTINED] }
        version: { type: integer, minimum: 0 }
        title: { type: string, maxLength: 200 }
        parentId: { type: string, format: uuid }
        sequenceNo: { type: integer, minimum: 1 }
        accessLevel: { type: string, enum: [FREE] }
        providerName: { type: string, const: YOUTUBE }
        youtubeVideoId: { type: string, pattern: '^[A-Za-z0-9_-]{11}$' }
        slug: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        summary: { type: string, maxLength: 4000 }
        hskLevel: { type: integer, minimum: 1, maximum: 9 }
        transcriptHanzi: { type: string, maxLength: 20000 }
        transcriptPinyin: { type: string, maxLength: 20000 }
        translationVi: { type: string, maxLength: 20000 }
        dictationHint: { type: string, maxLength: 4000 }
        startMilliseconds: { type: integer, minimum: 0 }
        endMilliseconds: { type: integer, minimum: 0 }
        segmentType: { type: string, enum: [DICTATION, SHADOWING, BOTH, PLAYBACK_ONLY] }
        mediaKind: { type: string, const: VIDEO }
        altText: { type: string, maxLength: 500 }
        approvalEligible: { type: boolean, description: Safe indication that PENDING_SCAN YouTube media may be approved after trusted-source verification. }
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    Meta:
      type: object
      required: [correlationId]
      properties: { correlationId: { type: string, format: uuid } }
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, enum: [VALIDATION_ERROR, UNAUTHENTICATED, AUTHORIZATION_DENIED, RESOURCE_NOT_FOUND, STATE_CONFLICT] }
        message: { type: string }
    ContentEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success: { const: true }
        data: { $ref: '#/components/schemas/ContentProjection' }
        error: { type: 'null' }
        meta: { $ref: '#/components/schemas/Meta' }
    ContentListEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success: { const: true }
        data: { type: array, items: { $ref: '#/components/schemas/ContentProjection' } }
        error: { type: 'null' }
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
            - $ref: '#/components/schemas/PageMeta'
    ErrorEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success: { const: false }
        data: { type: 'null' }
        error: { $ref: '#/components/schemas/Error' }
        meta: { $ref: '#/components/schemas/Meta' }
security:
  - bearerAuth: []